requirements.txt
fastapi==0.104.0
uvicorn==0.23.2
pydantic==2.4.2
sentence-transformers==2.2.2
faiss-cpu==1.7.4
google-generativeai==0.3.1
python-dotenv==1.0.0
python-multipart==0.0.6
aiofiles==23.2.1
jinja2==3.1.2
typing-extensions==4.8.0
numpy==1.26.0 